requirements.txt
aiohttp==3.9.5 ; python_version >= "3.12" and python_version < "3.13"
aiosignal==1.3.1 ; python_version >= "3.12" and python_version < "3.13"
alluka==0.3.1 ; python_version >= "3.12" and python_version < "3.13"
annotated-types==0.7.0 ; python_version >= "3.12" and python_version < "3.13"
anyio==4.4.0 ; python_version >= "3.12" and python_version < "3.13"
asgi-prometheus==1.1.2 ; python_version >= "3.12" and python_version < "3.13"
asgi-tools==0.76.0 ; python_version >= "3.12" and python_version < "3.13"
asgiref==3.8.1 ; python_version >= "3.12" and python_version < "3.13"
attrs==23.2.0 ; python_version >= "3.12" and python_version < "3.13"
blacksheep==2.0.6 ; python_version >= "3.12" and python_version < "3.13"
bloxlink-lib @ git+https://github.com/bloxlink/bloxlink-lib@32d44631bd00a1ad821a4ffb2f9adf72893f42b7 ; python_version >= "3.12" and python_version < "3.13"
certifi==2024.7.4 ; python_version >= "3.12" and python_version < "3.13"
cffi==1.16.0 ; python_version >= "3.12" and python_version < "3.13"
charset-normalizer==3.1.0 ; python_version >= "3.12" and python_version < "3.13"
click==8.1.7 ; python_version >= "3.12" and python_version < "3.13"
colorama==0.4.6 ; python_version >= "3.12" and python_version < "3.13" and (sys_platform == "win32" or platform_system == "Windows")
colorlog==6.8.2 ; python_version >= "3.12" and python_version < "3.13"
discord-py==2.4.0 ; python_version >= "3.12" and python_version < "3.13"
dnspython==2.6.1 ; python_version >= "3.12" and python_version < "3.13"
essentials-openapi==1.0.9 ; python_version >= "3.12" and python_version < "3.13"
essentials==1.1.5 ; python_version >= "3.12" and python_version < "3.13"
frozenlist==1.4.1 ; python_version >= "3.12" and python_version < "3.13"
guardpost==1.0.2 ; python_version >= "3.12" and python_version < "3.13"
h11==0.14.0 ; python_version >= "3.12" and python_version < "3.13"
hikari @ git+https://github.com/hikari-py/hikari.git@318a670c2687ea641e1727f83444dc782a5f1a09 ; python_version >= "3.12" and python_version < "3.13"
hikari-yuyo[asgi]==1.19.1 ; python_version >= "3.12" and python_version < "3.13"
hikari[server] @ git+https://github.com/hikari-py/hikari.git@318a670c2687ea641e1727f83444dc782a5f1a09 ; python_version >= "3.12" and python_version < "3.13"
hiredis==2.3.2 ; python_version >= "3.12" and python_version < "3.13"
http-router==4.1.2 ; python_version >= "3.12" and python_version < "3.13"
httptools==0.6.1 ; python_version >= "3.12" and python_version < "3.13"
humanize==4.10.0 ; python_version >= "3.12" and python_version < "3.13"
idna==3.7 ; python_version >= "3.12" and python_version < "3.13"
itsdangerous==2.1.2 ; python_version >= "3.12" and python_version < "3.13"
markdown-it-py==3.0.0 ; python_version >= "3.12" and python_version < "3.13"
markupsafe==2.1.5 ; python_version >= "3.12" and python_version < "3.13"
mdurl==0.1.2 ; python_version >= "3.12" and python_version < "3.13"
motor==3.3.2 ; python_version >= "3.12" and python_version < "3.13"
multidict==6.0.5 ; python_version >= "3.12" and python_version < "3.13"
prometheus-client==0.20.0 ; python_version >= "3.12" and python_version < "3.13"
psutil==5.9.0 ; python_version >= "3.12" and python_version < "3.13"
pycparser==2.22 ; python_version >= "3.12" and python_version < "3.13"
pydantic-core==2.20.1 ; python_version >= "3.12" and python_version < "3.13"
pydantic==2.8.2 ; python_version >= "3.12" and python_version < "3.13"
pygments==2.18.0 ; python_version >= "3.12" and python_version < "3.13"
pymongo==4.8.0 ; python_version >= "3.12" and python_version < "3.13"
pynacl==1.5.0 ; python_version >= "3.12" and python_version < "3.13"
python-dateutil==2.8.2 ; python_version >= "3.12" and python_version < "3.13"
python-dotenv==1.0.1 ; python_version >= "3.12" and python_version < "3.13"
pyyaml==6.0.1 ; python_version >= "3.12" and python_version < "3.13"
rapidfuzz==3.9.4 ; python_version >= "3.12" and python_version < "3.13"
redis==5.0.1 ; python_version >= "3.12" and python_version < "3.13"
requests==2.32.2 ; python_version >= "3.12" and python_version < "3.13"
rich==13.7.1 ; python_version >= "3.12" and python_version < "3.13"
rodi==2.0.6 ; python_version >= "3.12" and python_version < "3.13"
sentry-sdk==1.45.0 ; python_version >= "3.12" and python_version < "3.13"
six==1.16.0 ; python_version >= "3.12" and python_version < "3.13"
sniffio==1.3.1 ; python_version >= "3.12" and python_version < "3.13"
thefuzz==0.22.1 ; python_version >= "3.12" and python_version < "3.13"
typing-extensions==4.12.2 ; python_version >= "3.12" and python_version < "3.13"
urllib3==2.2.2 ; python_version >= "3.12" and python_version < "3.13"
uvicorn[standard]==0.27.1 ; python_version >= "3.12" and python_version < "3.13"
uvloop==0.19.0 ; (sys_platform != "win32" and sys_platform != "cygwin") and platform_python_implementation != "PyPy" and python_version >= "3.12" and python_version < "3.13"
watchfiles==0.22.0 ; python_version >= "3.12" and python_version < "3.13"
websockets==12.0 ; python_version >= "3.12" and python_version < "3.13"
yarl==1.9.4 ; python_version >= "3.12" and python_version < "3.13"